move components to SRC folder, adapt page imports
