+++ SRC/components/BagDetails.jsx
