+++ SRC/components/BagDetailsItem.jsx
