+++ pages/details/items/[slug].js
@@ -1,7 +1,7 @@
 import React from "react";
 import { getItem, getItems } from "../../../utils/data/getData";
-import ItemDetails from "../../../components/ItemDetails";
-import Layout from "../../../components/layouts/layout/Layout";
+import ItemDetails from "../../../SRC/components/ItemDetails";
+import Layout from "../../../SRC/components/layouts/layout/Layout";
 
 const Details = ({
   data: {
